git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/gpl-compat@26349 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/dvdplayer/DVDTSCorrection.cpp
@@ -90,7 +90,7 @@ void CPullupCorrection::Add(double pts)
     m_haspattern = true;
     m_patternlength = m_pattern.size();
     CLog::Log(LOGDEBUG, "CPullupCorrection: detected pattern of length %i: %s",
-              pattern.size(), GetPatternStr().c_str());
+              (int)pattern.size(), GetPatternStr().c_str());
   }
   
   //calculate where we are in the pattern
@@ -288,3 +288,4 @@ CStdString CPullupCorrection::GetPatternStr()
   
   return patternstr;
 }
+
